pgwire: retry on no complete timestamps error
Implement a retry loop in the coord client used by the pgwire session
to retry incomplete timestamp errors. These errors generally occur
because users are waiting on a newly created or initialized view to
process data. We don't know the first timestamp that will be produced,
so we can't use something like `AS OF now()`. Because the session itself
is retrying, it can avoid putting itself into a failed state.

Fixes #3122
Fixes #8163